backend/requirements.txt
Flask
flask-cors
flask-bcrypt
flask-pymongo
PyJWT
pymongo
python-dotenv